LaTeX resume: uses inverted Deedy template with custom spacing.
Download the Deedy package files in addition if you want to use this stylesheet. I flipped the two column version and added a stronger header with a horizontal bar, and color for the links. 
Build with LuaLatex, note that empty location{} or descrip{} (and the color tags) give warnings - I just ignore them because it builds correctly.